returns no longer display return date when returning. it was a testing feature that didn't get cleaned up. added comments to the file writer. customer menu no longer prints twice when returning a product. removed variable anotherproduct, since it was never used. (Line 1008.) If, for any reason, you want to return it, the specific line of code was                              Product anotherProduct = null;
removed currentproduct = null; on line 1027 because it's redundant.
removed readUUID in the end of the dart controller as it was never used.
just making the code look more organized and prettier.
removed leftover compareTo on line 114. if you want to return it, here's the code:
     public int compareTo(Product anotherProduct) {

        if(this.averageRating < anotherProduct.getReleaseYear()) {
            return -1;
        } else if (this.averageRating == anotherProduct.getReleaseYear()) {
            return 0;
        } else {
            return 1;
        }
    }
removed public Customer(){} on line 41 in Customer.java.

Sorting didn't work for games, and used to print albums along with games. Fixed.
Sorting now shows nonrated products at the end of the list.
added a halloween easter egg :D